Add files via upload
The purpose of the dashboard is to give insight to the economic status of South Africa. The dashboard focuses on a range of indicators and time periods are used to filter and see the change in time. It is broken down from the first half of the year 2023, Monthly performance, annual performance and quarterly performance from 2010 to 2023.